[Refactor #131] img svg 변경
