+++ style/livekit/livekit.module.css
@@ -44,8 +44,7 @@
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
-  left: 0;
-  top: 0;
+  inset: 0;
   z-index: 1;
   border-radius: var(--border-radius);
   cursor: pointer;
