+++ style/mainPage/main.module.css
@@ -52,7 +52,7 @@
   transform: translateY(-55%);
   width: 20px;
   height: 20px;
-  background: url("../../assets/images/icon_search.png") no-repeat;
+  background: url("../../assets/images/icon_search.svg") no-repeat;
   cursor: pointer;
 }
 
@@ -124,7 +124,7 @@
   left: 20px;
   top: 50%;
   transform: translateY(-51%);
-  background: url("../../assets/images/icon_point.png") no-repeat;
+  background: url("../../assets/images/icon_point.svg") no-repeat;
   width: 20px;
   height: 21px;
 }
@@ -157,19 +157,18 @@
 
 .gameGoButton button.makeRoom:before {
   content: "";
-  width: 15px;
-  height: 15px;
-  background: url("../../assets/images/icon_plus.png") no-repeat;
+  width: 20px;
+  height: 20px;
+  background: url("../../assets/images/icon_plus.svg") no-repeat;
   position: absolute;
-  left: 15px;
-  top: 50%;
+  left: 10px;
+  top: 45%;
   transform: translateY(-48%);
 }
 
 .roomList {
   display: flex;
-  gap: 80px;
-  justify-content: space-between;
+  gap: 125px;
   flex-wrap: wrap;
 }
 
@@ -250,14 +249,12 @@
   justify-content: center;
 }
 
-/*---------------- respansive --------------*/
-@media (max-width: 1780px) {
+/*---------------- Responsive --------------*/
+@media (max-width: 860px) {
   .roomList {
     justify-content: center;
   }
-}
 
-@media (max-width: 860px) {
   .MainGnb {
     flex-wrap: wrap;
   }
