+++ types/index.ts
@@ -45,6 +45,7 @@ export interface CountState {
   setTimer: (newCount: number) => void;
   setIsStart: (newToggle: boolean) => void;
 }
+
 export interface ReadyState {
   isReady: boolean;
   setIsReady: (newModal: boolean) => void;
